Update cookie and token revocation handling
We will ignore all cookies set by the backend and don't use them in any
of our requests.

Regarding the token revocation: As it turned out the refresh token can't
be revoked directly. The refresh token is linked to the access token, so
by revoking the access token, the refresh token is invalidated indirectly.

Closes #11